refactor; test
