+++ gamechangerml/src/entity/aux_data/entities.csv
@@ -205,7 +205,6 @@ Secretary of the Air Force,"",GCPER
 Secretary of the Army,"",GCPER
 Secretary of Defense,SecDef,GCPER
 Secretary of the Navy,"",GCPER
-Secretary of the Department of Homeland Security,"",GCPER
 Secretary of the Department of Veterans Affairs,"",GCPER
 Selective Service System,SSS,GCORG
 Sergeant Major of the Army,SMA,GCPER
